# src/resources/swagger/vc-http-api.yaml
openapi: 3.0.0
info:
  version: "0.0.0"
  title: VC HTTP API
  description: This is an Experimental Open API Specification for the [VC Data Model](https://www.w3.org/TR/vc-data-model/).
  license:
    name: W3C Software and Document License
    url: http://www.w3.org/Consortium/Legal/copyright-software.
  contact:
    name: GitHub Source Code
    url: https://github.com/w3c-ccg/vc-http-api
...
servers:
  - url: /services/
paths:
  /issue/credentials:
    post:
      tags:
        - Issuer
      summary: Issues a credential and returns it in the response body.
      operationId: issueCredential
      description: Issues a credential and returns it in the response body.
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueCredentialRequest"
        description: Parameters for issuing the credential.
      responses:
        "201":
          description: Credential successfully issued!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueCredentialResponse"
        "400":
          description: invalid input!
        "500":
          description: error!
  /issue:
    post:
      tags:
        - Issuer
      summary: Create a new issue configuration.
      description: Creates a new issue configuration under the authenticated user.
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssuerConfig"
      responses:
        '200':
          description: success!
        '400':
          description: invalid input!
        '500':
          description: error!
  /issue/{id}:
    get:
      tags:
        - Issuer
      summary: Get an issuer configuration by id.
      description: Gets an issuer configuration associated with the authenticated user
      parameters:
        - name: id
          in: path
          description: Issuer id
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: success!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssuerConfig"
        '400':
          description: invalid input!
        '500':
          description: error!
  /issue/{id}/credentials:
    post:
      tags:
        - Issuer
      summary: Issues a credential via the specified config.
      description: Issues a credential via the specified config parameters and returns it in the response body.
      parameters:
        - name: id
          in: path
          description: Issuer id
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComposeAndIssueRequest"
        description: Parameters for issuing the credential.
      responses:
        "201":
          description: Credential successfully issued!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueCredentialResponse"
        "400":
          description: invalid input!
        "500":
          description: error!
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  schemas:
    LinkedDataProof:
      type: object
      description: A JSON-LD Linked Data proof.
      properties:
        type:
          type: string
          description: Linked Data Signature Suite used to produce proof.
        created:
          type: string
          description: Date the proof was created.
        verificationMethod:
          type: string
          description: Verification Method used to verify proof.
        proofPurpose:
          type: string
          description: The purpose of the proof to be used with verificationMethod.
        jws:
          type: string
          description: Detached JSON Web Signature
      example:
        created: '2020-04-02T18:28:08Z'
        jws: >-
          eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..YtqjEYnFENT7fNW-COD0HAACxeuQxPKAmp4nIl8jYAu__6IH2FpSxv81w-l5PvE1og50tS9tH8WyXMlXyo45CA
        proofPurpose: assertionMethod
        type: Ed25519Signature2018
        verificationMethod: 'did:example:123#z6MksHh7qHWvybLg5QTPPdG2DgEjjduBDArV9EF9mRiRzMBN'
...
    Issuer:
      type: object
      description: A JSON-LD Verifiable Credential Issuer.
      oneOf:
        - type: string
        - type: object
          properties:
            id:
              type: string
      example:
        id: 'did:key:z6MkjRagNiMu91DduvCvgEsqLZDVzrJzFrwahc4tXLt9DoHd'
    CredentialStatus:
      type: object
      description: An object describing where to resolve the credential revocation status.
      properties:
        id:
          description: A URL that identifies the status information associated with the Verifiable Credential.
          type: string
        "type":
          description: The JSONLD object type. Currently RevocationList2020Status.
          type: string
          enum:
            - RevocationList2020Status
        revocationListIndex:
          description: >
            An arbitrary size integer greater than or equal to 0, expressed as a string.
            The value identifies the bit position of the revocation status of the verifiable credential.
          type: string
        revocationListCredential:
          description: A URL to the revocation list Verifiable Credential.
          type: string
    Credential:
      type: object
      description: A JSON-LD Verifiable Credential without a proof.
      properties:
        "@context":
          type: array
          description: The JSON-LD context of the credential.
        "credentialStatus":
          $ref: '#/components/schemas/CredentialStatus'
        "id":
          type: string
          description: The ID of the credential.
        "type":
          type: array
          description: The JSON-LD type of the credential.
        "issuer":
          $ref: "#/components/schemas/Issuer"
        "issuanceDate":
          type: string
          description: The issuanceDate
        "expirationDate":
          type: string
          description: The expirationDate
        "credentialSubject":
          type: object
          description: The subject
      example:
        '@context':
          - 'https://www.w3.org/2018/credentials/v1'
          - 'https://www.w3.org/2018/credentials/examples/v1'
          - 'https://w3id.org/vc-revocation-list-2020/v1'
        "credentialStatus":
          "id": "http://example.gov/credentials/3732#2"
          "type": "RevocationList2020Status"
          "revocationListIndex": "2"
          "revocationListCredential": "https://example.github.io/example-repo/revocation-credential.jsonld"
        credentialSubject:
          degree:
            name: Bachelor of Science and Arts
            type: BachelorDegree
          id: 'did:example:123'
        id: 'http://example.gov/credentials/3732'
        issuanceDate: '2020-03-16T22:37:26.544Z'
        issuer: 'did:example:123'
        type:
          - VerifiableCredential
          - UniversityDegreeCredential
    VerifiableCredential:
      type: object
      description: A JSON-LD Verifiable Credential with a proof.
      allOf:
        - $ref: "#/components/schemas/Credential"
        - type: object
          properties:
            proof:
              $ref: "#/components/schemas/LinkedDataProof"
      example:
        '@context':
          - 'https://www.w3.org/2018/credentials/v1'
          - 'https://www.w3.org/2018/credentials/examples/v1'
        credentialSubject:
          degree:
            name: Bachelor of Science and Arts
            type: BachelorDegree
          id: 'did:example:123'
        id: 'http://example.gov/credentials/3732'
        issuanceDate: '2020-03-16T22:37:26.544Z'
        issuer: 'did:example:123'
        proof:
          created: '2020-04-02T18:28:08Z'
          jws: >-
            eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..YtqjEYnFENT7fNW-COD0HAACxeuQxPKAmp4nIl8jYAu__6IH2FpSxv81w-l5PvE1og50tS9tH8WyXMlXyo45CA
          proofPurpose: assertionMethod
          type: Ed25519Signature2018
          verificationMethod: 'did:example:123#z6MksHh7qHWvybLg5QTPPdG2DgEjjduBDArV9EF9mRiRzMBN'
        type:
          - VerifiableCredential
          - UniversityDegreeCredential
...
    IssueCredentialRequest:
      type: object
      properties:
        credential:
          $ref: "#/components/schemas/Credential"
    IssueCredentialResponse:
      $ref: "#/components/schemas/VerifiableCredential"
...
    IssuerConfig:
      additionalProperties: false
      description: Configuration options for issuing credentials
      example:
        id: example-issuer-config
        context:
          - 'https://www.w3.org/2018/credentials/examples/v1'
        type:
          - UniversityDegreeCredential
        revocationListCredential: "https://example.github.io/example-repo/revocation-credential.jsonld"
      properties:
        id:
          description: The id used to reference the configuration.
          type: string
        contexts:
          description: >
            An array of contexts to be used in the issued credential. If not included, the W3C VC context will be added
            as the first value (https://www.w3.org/2018/credentials/v1). If revocationListCredential is specified, then
            the RevocationList2020 context will also be automatically added if not already present
            (https://w3id.org/vc-revocation-list-2020/v1).
          type: array
        type:
          description: >
            The JSONLD type objects used in the credential. They must be defined in the added contexts. If
            "VerifiableCredential" is not the first type, it will be automatically added.
          type: array
        revocationListCredential:
          description: A URL pointing to the revocationListCredential to be used for revoking credentials.
          type: string
    ComposeAndIssueRequest:
      additionalProperties: false
      description: Properties required to issue a credential with an issuer config
      example:
        credentialSubject:
            degree:
              name: Bachelor of Science and Arts
              type: BachelorDegree
            id: "did:example:123"
        revocationListIndex: "2"
      properties:
        revocationListIndex:
          description: >
            An arbitrary size integer greater than or equal to 0, expressed as a string.
            The value identifies the bit position of the revocation status of the verifiable credential.
          type: string